replication: fix transaction boundary errors during upgrade
Upgrading cluster from versions that do not have transaction boundary
support (say, 1.10) has already caused some inconvenience:
9fcbbb3e7d5e9f5a876ee27a7bf93303321e26b2 ("applier: fix tx boundary check for
half-applied txns")

Now there's a similar issue. Imagine the cluster is undergoing an
upgrade, and there are some old and some new instances.
Let's take a look at 3 particular instances, namely, A, B and C.
Imagine A and C are new, and B is old.

A is writable, so it constantly generates new transactions and
replicates them to the peers.

Let's also imagine that C is just started, so it tries to establish
connection to both B and A.

Now, B is old, so it doesn't preserve transaction boundaries. And the
transaction coming from A
{[row1: lsn=1, tsn=1], [row2: lsn=2, tsn=1, is_commit=true]}
is written to its journal simply as a pair of single-statement
transactions:
{[row1: lsn=1]}, {[row2: lsn=2]}.

It's possible that C first connects to B, and receives part of the
transaction, namely, {[row1: lsn=1]}. Now for some reason applier to B
yields, and the node is connected to A. A sends C what's missing:
[row2: lsn=2, tsn=1, is_commit=true]. This packet already has
transaction boundaries, so it triggers transaction boundary check on C,
which's obviously violated: row lsn (2) isn't equal to row tsn (1), and
it's the first row that came from this connection. So it should surely
be the beginning of the transaction. This makes C's applier to A stop
with the following error:

ER_PROTOCOL: Transaction id must be equal to LSN of the first row in the
transaction.

Now, even if the user tries to restart replication on C, he might
succeed, or he might fail: it's possible that the same situation when B
relays part of the transaction to C before A relays the rest repeats
over and over again.

In a big enough cluster, when part of the nodes is upgraded and part is
not, and each node is writable and generates new transactions, it's
almost impossible to catch the correct moment in time to restart
replication. The cluster will almost surely descend to a state where
all upgraded nodes are disconnected from each other due to transaction
boundary errors.

Interestingly, the new nodes will remain connected via the old nodes,
since old nodes clear the replication stream from all the transaction
boundaries, so there are no errors when replicating from an old node.